No more magic timeouts

+++ util.go
@@ -9,7 +9,7 @@ import (
 // the network.  This is based on the number of AC zero crossings, the
 // number of times a message can be repeated (ttl) and whether or not
 // the message is an extended message
-func PropagationDelay(ttl int, extended bool) (pd time.Duration) {
+func PropagationDelay(ttl uint8, extended bool) (pd time.Duration) {
 	// wait 2 * ttl * message length zero crossings
 	if extended {
 		pd = time.Second * 26 * time.Duration(ttl) / 60
